requirements.txt
groq>=0.9.0
psycopg2-binary>=2.9.9
